[GEOT-4284] infinite loop using uom w/ GraphicStroke patch by Akos Maroy

+++ modules/library/render/src/main/java/org/geotools/renderer/style/MarkStyle2D.java
@@ -38,7 +38,7 @@ public class MarkStyle2D extends PolygonStyle2D {
     static boolean maxMarkSizeEnabled = Boolean.getBoolean("org.geotools.maxMarkSizeEnabled");
     
     Shape shape;
-    int size;
+    double size;
     float rotation;
 
     /**
@@ -104,7 +104,7 @@ public class MarkStyle2D extends PolygonStyle2D {
      * Returns the size of the shape, in pixels
      *
      */
-    public int getSize() {
+    public double getSize() {
         return size;
     }
 
@@ -131,7 +131,7 @@ public class MarkStyle2D extends PolygonStyle2D {
      *
      * @param i
      */
-    public void setSize(int i) {
+    public void setSize(double i) {
         size = i;
     }
